+++ CHANGELOG.md
@@ -6,9 +6,6 @@ From version 2.6.0, the sections in this file adhere to the [keep a changelog](h
 * [#2188](https://github.com/Shopify/shopify-cli/pull/2188): Update URLs by default on serve and add --no-update flag to skip it
 * [#2203](https://github.com/Shopify/shopify-cli/pull/2203): Use javy version 0.3.0
 
-### Added
-* [#2200](https://github.com/Shopify/shopify-cli/pull/2200): Add `theme share` command
-
 ### Fixed
 * [#2162](https://github.com/Shopify/shopify-cli/pull/2162): Improve encoding error handling for Checkout Extension localization
 * [#2187](https://github.com/Shopify/shopify-cli/pull/2187): Fix app serve after rails update
@@ -18,6 +15,7 @@ From version 2.6.0, the sections in this file adhere to the [keep a changelog](h
 
 ### Added
 * [#2190](https://github.com/Shopify/shopify-cli/pull/2190): Better login experience with spinner
+* [#2200](https://github.com/Shopify/shopify-cli/pull/2200): Add `theme share` command
 
 ## Version 2.15.2